Add setting default values for CausationCategory and DefectType when a new Defect is added. Add safety check that these default values cannot be deleted from database

+++ src/main/java/com/example/defecttrackerserver/core/defect/defectType/DefectTypeServiceImpl.java
@@ -77,6 +77,10 @@ public class DefectTypeServiceImpl implements DefectTypeService {
         if(!defectRepository.findByDefectTypeId(id).isEmpty())
             throw new UnsupportedOperationException("DefectType cannot be deleted because it is used in Defects");
 
+        if(defectType.getName().equals("Undefined"))
+            throw new UnsupportedOperationException("DefectType 'Undefined' cannot be deleted " +
+                    "because it is a default value.");
+
         defectTypeRepository.delete(defectType);
     }
